Fix DevicePool 32-bit offset wraparound; skip Hausdorff in beam loop; add get_part_info API
- Widen DevicePool offset/capacity from unsigned int to unsigned long long.
  480 concurrent expansion blocks each allocating ~15MB D&C scratch exceeded
  4GB, causing silent unsigned wraparound → overlapping memory → corrupted
  edge pool free lists → illegal memory access in btpool_new (confirmed via
  compute-sanitizer on Octocat-v2 at block 457).
- Remove 4GB scratch cap (no longer needed with 64-bit offset).
- Skip Hausdorff kernel in beam loop for now (Rv-only termination).
- Add beam_get_part_info C API + Python wrapper for per-part diagnostics
  (rv_cost, hausdorff, mesh_volume, hull_volume).

+++ cuda/beam_v2.cu
@@ -13,8 +13,8 @@
 __device__ inline void* global_alloc_warp(DevicePool* gpool, int bytes, int lane) {
     long long ptr_ll = 0LL;
     if (lane == 0) {
-        unsigned int aligned = ((unsigned int)bytes + 15) & ~15;
-        unsigned int old = atomicAdd(gpool->offset, aligned);
+        unsigned long long aligned = (unsigned long long)((bytes + 15) & ~15);
+        unsigned long long old = atomicAdd(gpool->offset, aligned);
         if (old + aligned <= gpool->capacity)
             ptr_ll = (long long)(gpool->base + old);
     }
@@ -24,8 +24,8 @@ __device__ inline void* global_alloc_warp(DevicePool* gpool, int bytes, int lane
 
 // Helper: thread-0-only alloc from DevicePool (for block-level use, NOT warp)
 __device__ inline void* global_alloc_t0(DevicePool* gpool, int bytes) {
-    unsigned int aligned = ((unsigned int)bytes + 15) & ~15;
-    unsigned int old = atomicAdd(gpool->offset, aligned);
+    unsigned long long aligned = (unsigned long long)((bytes + 15) & ~15);
+    unsigned long long old = atomicAdd(gpool->offset, aligned);
     if (old + aligned > gpool->capacity) return NULL;
     return gpool->base + old;
 }
